+++ date.py
@@ -1,0 +1,17 @@
+import pyttsx3
+import datetime
+import wikipedia
+import webbrowser as wb
+import psutil
+import pyjokes
+import pyautogui
+from speak immport speak
+
+def date_():
+    year = datetime.datetime.now().year
+    month = datetime.datetime.now().month
+    date = datetime.datetime.now().day
+    speak("Today's date is")
+    speak(date)
+    speak(month)
+    speak(year)
