+++ speak.py
@@ -1,0 +1,13 @@
+import pyttsx3
+import datetime
+import wikipedia
+import webbrowser as wb
+import psutil
+import pyjokes
+import pyautogui
+from speak immport speak
+
+
+def speak(audio):
+    engine.say(audio)
+    engine.runAndWait()
