+++ time.py
@@ -1,0 +1,13 @@
+import pyttsx3
+import datetime
+import wikipedia
+import webbrowser as wb
+import psutil
+import pyjokes
+import pyautogui
+from speak immport speak
+
+def time_():
+    Time = datetime.datetime.now().strftime("%H:%M:%S")
+    speak("The current time is")
+    speak(Time)
